+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 **Known Issue:** this version was compiled using typescript 3.7!
 The resulting `.d.ts` are **only** compatible with typescript versions `>= 3.6`. See [typescript 3.7 release notes](www.typescriptlang.org/docs/handbook/release-notes/typescript-3-7.html#class-field-mitigations) for more information.
 
+### Added
+
+- Added `currentZoomTransform` to get the current zoom transform of the graph
+- Added `zoomToBox` to zoom to a specific area of the graph
+- Added 'zoomchange' events to the grapheditor
+- Added `getGraphPointFromClientCoordinates` as inverse of `getClientPointFromGraphCoordinates`
+
 ### Fixed
 
 - Fixed wrong imports depending on 'd3' instead of 'd3-selection'